Extract OAuth Client form as hook
ref DEV-1754
