+++ portal/src/graphql/portal/EditOAuthClientScreen.tsx
@@ -6,8 +6,6 @@ import {
   useParams,
   useSearchParams,
 } from "react-router-dom";
-import deepEqual from "deep-equal";
-import { produce, createDraft } from "immer";
 import {
   Icon,
   Text,
@@ -25,22 +23,9 @@ import ScreenContent from "../../ScreenContent";
 import NavBreadcrumb, { BreadcrumbItem } from "../../NavBreadcrumb";
 import ShowError from "../../ShowError";
 import ShowLoading from "../../ShowLoading";
-import EditOAuthClientForm, {
-  getReducedClientConfig,
-} from "./EditOAuthClientForm";
-import {
-  ApplicationType,
-  OAuthClientConfig,
-  OAuthClientSecret,
-  PortalAPIAppConfig,
-  PortalAPISecretConfig,
-  PortalAPISecretConfigUpdateInstruction,
-} from "../../types";
-import { clearEmptyObject } from "../../util/misc";
-import {
-  AppSecretConfigFormModel,
-  useAppSecretConfigForm,
-} from "../../hook/useAppSecretConfigForm";
+import EditOAuthClientForm from "./EditOAuthClientForm";
+import { ApplicationType, OAuthClientConfig } from "../../types";
+import { AppSecretConfigFormModel } from "../../hook/useAppSecretConfigForm";
 import FormContainer from "../../FormContainer";
 import ScreenLayoutScrollView from "../../ScreenLayoutScrollView";
 import styles from "./EditOAuthClientScreen.module.css";
@@ -57,6 +42,7 @@ import { AppSecretKey } from "./globalTypes.generated";
 import { startReauthentication } from "./Authenticated";
 import { useLocationEffect } from "../../hook/useLocationEffect";
 import { useAppSecretVisitToken } from "./mutations/generateAppSecretVisitTokenMutation";
+import { useOAuthClientForm } from "../../hook/useOAuthClientForm";
 
 interface FormState {
   publicOrigin: string;
@@ -76,90 +62,6 @@ function isLocationState(raw: unknown): raw is LocationState {
     typeof raw === "object" &&
     (raw as Partial<LocationState>).isClientSecretRevealed != null
   );
-}
-
-function constructFormState(
-  config: PortalAPIAppConfig,
-  secrets: PortalAPISecretConfig
-): FormState {
-  const clientSecretMap: Partial<Record<string, string>> =
-    secrets.oauthClientSecrets?.reduce<Record<string, string>>(
-      (acc: Record<string, string>, currValue: OAuthClientSecret) => {
-        if (currValue.keys?.length && currValue.keys.length >= 1) {
-          acc[currValue.clientID] = currValue.keys[0].key;
-        }
-        return acc;
-      },
-      {}
-    ) ?? {};
-  return {
-    publicOrigin: config.http?.public_origin ?? "",
-    clients: config.oauth?.clients ?? [],
-    editedClient: null,
-    removeClientByID: undefined,
-    clientSecretMap,
-  };
-}
-
-function constructConfig(
-  config: PortalAPIAppConfig,
-  secrets: PortalAPISecretConfig,
-  _initialState: FormState,
-  currentState: FormState,
-  _effectiveConfig: PortalAPIAppConfig
-): [PortalAPIAppConfig, PortalAPISecretConfig] {
-  const newConfig = produce(config, (config) => {
-    config.oauth ??= {};
-    config.oauth.clients = currentState.clients.slice();
-
-    if (currentState.removeClientByID) {
-      config.oauth.clients = config.oauth.clients.filter(
-        (c) => c.client_id !== currentState.removeClientByID
-      );
-      clearEmptyObject(config);
-      return;
-    }
-
-    const client = currentState.editedClient;
-    if (client) {
-      const index = config.oauth.clients.findIndex(
-        (c) => c.client_id === client.client_id
-      );
-      if (
-        index !== -1 &&
-        !deepEqual(
-          getReducedClientConfig(client),
-          getReducedClientConfig(config.oauth.clients[index]),
-          { strict: true }
-        )
-      ) {
-        config.oauth.clients[index] = createDraft(client);
-      }
-    }
-    clearEmptyObject(config);
-  });
-  return [newConfig, secrets];
-}
-
-function constructSecretUpdateInstruction(
-  _config: PortalAPIAppConfig,
-  _secrets: PortalAPISecretConfig,
-  currentState: FormState
-): PortalAPISecretConfigUpdateInstruction | undefined {
-  if (currentState.removeClientByID) {
-    return {
-      oauthClientSecrets: {
-        action: "cleanup",
-        cleanupData: {
-          keepClientIDs: currentState.clients
-            .filter((c) => c.client_id !== currentState.removeClientByID)
-            .map((c) => c.client_id),
-        },
-      },
-    };
-  }
-
-  return undefined;
 }
 
 interface FrameworkItem {
@@ -572,13 +474,7 @@ const EditOAuthClientScreen1: React.VFC<{
   secretToken: string | null;
 }> = function EditOAuthClientScreen1({ appID, clientID, secretToken }) {
   const { renderToString } = useContext(Context);
-  const form = useAppSecretConfigForm({
-    appID,
-    secretVisitToken: secretToken,
-    constructFormState,
-    constructConfig,
-    constructSecretUpdateInstruction,
-  });
+  const form = useOAuthClientForm(appID, secretToken);
   const { setState, save, isUpdating } = form;
 
   const featureConfig = useAppFeatureConfigQuery(appID);
